Add account media prune option to vacuum scheduler
This adds a new vacuum scheduler that will remove cached avatar images
and header images from accounts that have not sent an activity in the
specified number of days and are not following or followed by any users
on the server.

This may probably lead to missing avatar images from boosted posts of
users that are not following or followed by any users on the server.

The default setting is blank, which will effectively disable the vacuum
schedule for this new functionality.

Mastodon will re-fetch the profile data as new activities with the
associated actor come in.

TODO: Translations. En and De are by me, others have to be translated
by the translation team.